requirements.txt
torchvision 
torch
requests
pytests